requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
cached-property==1.5.1
coverage==4.5.4
decorator==4.4.0
Django==2.2.5
django-cacheops==4.2
django-extensions==2.2.1
django-redis==4.10.0
django-rest-framework==0.1.0
django-url-filter==0.3.13
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-renderer-xlsx==0.3.3
dry-rest-permissions==0.1.10
entrypoints==0.3
enum-compat==0.0.2
et-xmlfile==1.0.1
factory-boy==2.12.0
Faker==2.0.2
flake8==3.7.8
funcy==1.13
importlib-metadata==0.23
ipython==7.8.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.15.1
mccabe==0.6.1
more-itertools==7.2.0
openpyxl==2.6.3
packaging==19.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.0
prompt-toolkit==2.0.9
psycopg2-binary==2.8.3
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==1.6
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.1.2
pytest-cov==2.7.1
pytest-django==3.5.1
python-dateutil==2.8.0
pytz==2019.2
redis==3.3.8
six==1.12.0
sqlformatter==1.3
sqlparse==0.1.11
text-unidecode==1.3
traitlets==4.3.2
wcwidth==0.1.7
zipp==0.6.0
uwsgi==2.0.18
